+++ server/netlify.toml
@@ -1,0 +1,9 @@
+[functions]
+	external_node_modules = ["express"]
+	node_bundler = "esbuild"
+
+[[redirects]]
+	force = true
+	from = "/*"
+	status = 200
+	to = "/.netlify/functions/server/:splat"
