+++ server/package.json
@@ -14,9 +14,10 @@
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "ejs": "^2.7.4",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "mongoose": "^5.9.7",
     "request": "^2.88.2",
+    "serverless-http": "^3.2.0",
     "socket.io": "^2.3.0",
     "unirest": "^0.6.0"
   },
